Move ampnet to experimental
