+++ benchmarks/ampnet.py
@@ -20,7 +20,8 @@ from torchtext.data.utils import get_tokenizer
 from fairscale.nn import Pipe
 from fairscale.nn.model_parallel import initialize_model_parallel
 from fairscale.nn.model_parallel.initialize import get_data_parallel_group, get_pipeline_parallel_group
-from fairscale.nn.pipe import LazyModule, pipe
+from fairscale.nn.pipe import LazyModule
+from experimental.nn.pipe import pipe
 from fairscale.optim import GradScaler
 from fairscale.optim.oss import OSS
 from fairscale.utils.testing import dist_init, get_worker_map
@@ -508,7 +509,7 @@ def run_mp_worker(args, available_workers):
     model = blob["model"]
 
     balance = generate_balance(get_pipeline_parallel_group().size(), len(model))
-    p = pipe.Pipe(
+    p = pipe.AMPnetPipe(
         model,
         balance,
         style=Pipe.AsyncSchedule,
