+++ experimental/nn/pipe/ampnet.py
@@ -16,12 +16,12 @@ from torch.utils.data import DataLoader
 
 from fairscale.nn.model_parallel import get_pipeline_parallel_ranks
 
-from .async_schedule import AsyncMessageBody, AsyncMessageType, AsyncRecvOperator, Location, ModuleWrapper
-from .checkpoint import Checkpointing
-from .messages import Transport
-from .microbatch import Batch
-from .types import EVENT_LOOP_ACTIVATIONS_QUEUE, EVENT_LOOP_GRADIENTS_QUEUE, PipeMessage, TensorOrTensors
-from .worker import Task
+from fairscale.nn.pipe.async_schedule import AsyncMessageBody, AsyncMessageType, AsyncRecvOperator, Location, ModuleWrapper
+from fairscale.nn.pipe.checkpoint import Checkpointing
+from fairscale.nn.pipe.messages import Transport
+from fairscale.nn.pipe.microbatch import Batch
+from fairscale.nn.pipe.types import EVENT_LOOP_ACTIVATIONS_QUEUE, EVENT_LOOP_GRADIENTS_QUEUE, PipeMessage, TensorOrTensors
+from fairscale.nn.pipe.worker import Task
 
 
 def create_task_without_skip_trackers(
